API integration, bug found in Adding route and Adding schedule

+++ backend/ReservationBus/src/main/java/com/busBooking/ReservationBus/repos/BusRouteRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.awt.*;
 import java.util.Optional;
 
-public interface BusRouteRepository extends JpaRepository<BusRoute, Integer> {
+public interface BusRouteRepository extends JpaRepository<BusRoute, Long> {
 
    Optional<BusRoute> findByRouteName(String routeName);
    Optional<BusRoute> findByCityFromAndCityTo(String cityFrom, String cityTo);
